refactor(api): converge time + webhook interpreters into one default interpreter
Merge TimeInterpreter and WebhookInterpreter into a single DefaultInterpreter
that builds the agent-run input from (automation, triggerEvent). A webhook fire
renders the inbound headers+body as run context and tags the run with the
automation; a time fire is instruction-only, attaches the recurrence reschedule
callback, and tags the run with the schedule. Both paths attach the chat
callback so the run renders as a web-chat turn.

Rewire runScheduleNow$ (zero-schedules.service.ts) and dispatchAutomationWebhook$
(webhooks-automation.service.ts) to the single interpreter, add a
webhookRowToAutomation mapper as the counterpart to scheduleToAutomation, and
remove the now-redundant time-interpreter.ts / webhook-interpreter.ts split.
interpreterKind stays on the Automation as the future registry hook (one impl
for now). Behavior-preserving: existing schedule + webhook tests stay green
unchanged.

Part of #16847 (Slice 1). Closes #16850.
